fix: compilation issues
