--- SherpaFrontEnd/Dtos/Response.cs
@@ -1,5 +1,0 @@
-namespace SherpaFrontEnd.Dtos;
-
-public class Response
-{
-}
