--- SherpaFrontEnd/Dtos/Status.cs
@@ -1,6 +1,0 @@
-namespace SherpaFrontEnd.Dtos;
-
-public enum Status
-{
-    Draft
-}
