--- SherpaFrontEnd/Dtos/User.cs
@@ -1,13 +1,0 @@
-namespace SherpaFrontEnd.Dtos;
-
-public class User
-{
-    public Guid Id {get; set;}
-    public string Name {get; set;}
-
-    public User(Guid id, string name)
-    {
-        Id = id;
-        Name = name;
-    }
-}
