+++ SherpaFrontEnd/Pages/SurveyTable.razor
@@ -160,7 +160,7 @@
                 </th>
                 <td class="px-6 py-4">
                     <div class="flex items-center">
-                        <div class="h-2.5 w-2.5 rounded-full bg-red-500 mr-2"></div> @survey.Deadline
+                        <div class="h-2.5 w-2.5 rounded-full bg-red-500 mr-2"></div> @survey.Deadline.Date.ToString("dd/MM/yyyy")
                     </div>
                 </td>
                 <td class="px-6 py-4">
